Addressed Copilot review feedback on PR #4656
* Summed resident stubs in 64 bits so the per-run guard cannot wrap past ~505 runs
* Made the rescore no-op arm honor DiagnosticsOnly, so a diagnostics run whose
  first-pass key drifted folds the report instead of rewriting reconciled parquets
* Reworded the retired-switch startup error and the architecture doc to say only
  that the switch can no longer select the resident arm
* Named the actual retained-summary writer in the format-bump note

+++ pwiz_tools/Osprey/Osprey.Tasks/FirstPassFdrTask.cs
@@ -233,7 +233,7 @@ namespace pwiz.Osprey.Tasks
 
             // The analysis-wide retained base_id summary is deliberately NOT declared here,
             // unlike its experiment-sidecar sibling above. Declaring it would oblige
-            // WriteRetainedBaseIds to stamp it too, and until it has been stamped once,
+            // WriteRetainedBaseIdSummary to stamp it too, and until it has been stamped once,
             // OnlyDiagnosticsProductOutstanding reads every completed analysis on disk as owing
             // a first pass - so `--task ModelDiagnostics` on a finished 446-run cohort would
             // re-run Stage 1-5 for hours instead of folding the report in seconds, which is the
@@ -1118,7 +1118,9 @@ namespace pwiz.Osprey.Tasks
             // its report over the pre-compaction pool. A wrong report, silently, at MORE memory
             // than the route this arm replaced. Fail loudly instead. Routing the tokened case
             // to the overlay is the eventual answer; it waits on a gate that exercises a token.
-            int residentStubs = 0;
+            // 64-bit like the sibling sum in LoadOwnReconciliationBundle: at ~4.2 M stubs a run
+            // an int wraps negative past ~505 runs, and a negative count would pass this check.
+            long residentStubs = 0;
             foreach (var kvp in perFileEntries)
                 residentStubs += kvp.Value.Count;
             if (residentStubs > 0)
